create new files for split input
