+++ Makefile
@@ -22,7 +22,8 @@ LIBS	:= $(LKLIBFT) -lreadline
 
 SRCS	:=	./src/main.c \
 			./src/handle_env_var.c \
-			./src/trans_env.c
+			./src/trans_env.c \
+			./src/split_input.c
 			
 OBJS	:= ${SRCS:.c=.o}
